requirements.txt
bson==0.5.10
click==7.1.2
DateTime==4.3
dnspython==2.0.0
Flask==1.1.2
Flask-Cors==3.0.9
Flask-PyMongo==2.3.0
gunicorn==20.0.4
itsdangerous==1.1.0
jinja2==2.11.3
MarkupSafe==1.1.1
pymongo==3.11.0
python-dateutil==2.8.1
python-dotenv~=0.15.0
pytz==2020.1
six==1.15.0
Werkzeug==1.0.1
zope.interface==5.1.2
oauthlib~=3.1.0
pyOpenSSL==19.0.0
Flask-Login==0.4.1
pip~=21.0.1
wheel~=0.34.2
lxml~=4.6.2
setuptools~=47.1.0